fix styling and report creation

+++ frontend/src/routes/(app)/report/[owner]/[report]/+page.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import { invalidate } from '$app/navigation';
 	import Element from '$lib/components/report/Element.svelte';
-	import { ReportElementType, ZenoService } from '$lib/zenoapi';
+	import { ReportElementType, ZenoService, type ReportElement } from '$lib/zenoapi';
 	import Button, { Label } from '@smui/button';
 
 	export let data;
@@ -12,6 +12,7 @@
 	$: elements = data.reportElements;
 
 	function deleteElement(elementIndex: number) {
+		if (elementIndex < 0) return;
 		ZenoService.deleteReportElement(elementIndex).then(() => invalidate('app:report'));
 	}
 
@@ -23,31 +24,44 @@
 		}).then(() => invalidate('app:report'));
 	}
 
-	// function updateElement(event: CustomEvent<{ element: ReportElement }>) {
-	// 	ZenoService.updateReportElement(event.element).then(() => {
-	// 		if ($report) elementRequest = ZenoService.getReportElements($report.id);
-	// 	});
-	// }
+	function updateElement(event: CustomEvent<{ element: ReportElement }>) {
+		ZenoService.updateReportElement(report.id, event.detail.element as ReportElement).then(() =>
+			invalidate('app:report')
+		);
+	}
 </script>
 
-<div class="flex flex-col ml-12 h-full">
-	{#if report.editor}
-		<Button
-			style="width: 24px; height: 24px;background-color:var(--G5);position:absolute;right:50px;top:10px"
-			on:mouseleave={blur}
-			on:focusout={blur}
-			on:click={() => (isEdit = !isEdit)}
-		>
-			<Label>{isEdit ? 'View' : 'Edit'}</Label>
-		</Button>
-	{/if}
-	<h1>{report.name}</h1>
-	<div class="flex flex-col overflow-y-auto py-5">
-		{#each elements as element}
-			<Element {element} {isEdit} />
-		{/each}
-		{#if isEdit}
-			<button on:click={() => addElement(0)}>Add Element</button>
-		{/if}
+<div class="w-full bg-primary-light">
+	<div class="flex flex-col h-full max-w-4xl m-auto bg-background px-10">
+		<div class="flex items-center mt-12">
+			<h1 class="text-4xl mr-6" contenteditable={isEdit ? true : false}>
+				{report.name}
+			</h1>
+			{#if report.editor}
+				<Button
+					variant="raised"
+					on:mouseleave={blur}
+					on:focusout={blur}
+					on:click={() => (isEdit = !isEdit)}
+				>
+					<Label>{isEdit ? 'View' : 'Edit'}</Label>
+				</Button>
+			{/if}
+		</div>
+		<div class="flex flex-col overflow-y-auto py-5">
+			{#each elements as element}
+				<Element
+					{element}
+					{isEdit}
+					on:update={updateElement}
+					on:delete={() => deleteElement(element.id ?? -1)}
+				/>
+			{/each}
+			{#if isEdit}
+				<Button style="background-color:var(--G5);" on:click={() => addElement(0)}
+					>Add Element</Button
+				>
+			{/if}
+		</div>
 	</div>
 </div>
